# shared/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: VoxArena API
  version: "0.1.0"
  description: >
    Server-authoritative API for VoxArena (competitive singing game).
    Reads are public; write endpoints require authentication. Identity comes
    from a Supabase access token (`Authorization: Bearer <token>`) in
    production, or — only when `AUTH_DEV_BYPASS=true` and not in production —
    from an `x-player-id` header for local development.
servers:
  - url: http://localhost:3000
    description: Local dev
...
paths:
...
  /store/packs:
    get:
      tags: [Store]
      summary: List purchasable song packs
      description: Public; includes `owned` per pack when authenticated.
      security:
        - {}
        - bearerAuth: []
        - devPlayerId: []
      responses:
        "200":
          description: Active packs
          content:
            application/json:
              schema:
                type: object
                properties:
                  packs:
                    type: array
                    items: { $ref: "#/components/schemas/Pack" }
  /store/checkout:
    post:
      tags: [Store]
      summary: Start a Stripe Checkout session for a pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packId]
              properties:
                packId: { type: string, format: uuid }
      responses:
        "201":
          description: Checkout session URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  sessionId: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Not purchasable yet, or already owned }
        "503": { description: Payments not configured (STRIPE_SECRET_KEY) }
...
  /cosmetics:
    get:
      tags: [Cosmetics]
      summary: List active cosmetic items
      description: Public; includes `owned` and `equipped` per item when authenticated.
      security:
        - {}
        - bearerAuth: []
        - devPlayerId: []
      responses:
        "200":
          description: Active cosmetics
          content:
            application/json:
              schema:
                type: object
                properties:
                  cosmetics:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        slug: { type: string }
                        name: { type: string }
                        description: { type: string, nullable: true }
                        category: { type: string }
                        priceCents: { type: integer }
                        currency: { type: string }
                        purchasable: { type: boolean }
                        owned: { type: boolean }
                        equipped: { type: boolean }
  /cosmetics/checkout:
    post:
      tags: [Cosmetics]
      summary: Start a Stripe Checkout session for a cosmetic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cosmeticItemId]
              properties:
                cosmeticItemId: { type: string, format: uuid }
      responses:
        "201":
          description: Checkout session URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  sessionId: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Not purchasable yet, or already owned }
        "503": { description: Payments not configured (STRIPE_SECRET_KEY) }
  /cosmetics/equip:
    post:
      tags: [Cosmetics]
      summary: Equip an owned cosmetic (unequips others in its category)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cosmeticItemId]
              properties:
                cosmeticItemId: { type: string, format: uuid }
      responses:
        "200":
          description: "{ equipped, category }"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: You do not own this cosmetic }
  /cosmetics/unequip:
    post:
      tags: [Cosmetics]
      summary: Unequip an owned cosmetic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cosmeticItemId]
              properties:
                cosmeticItemId: { type: string, format: uuid }
      responses:
        "200":
          description: "{ unequipped: boolean }"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase access token
    devPlayerId:
      type: apiKey
      in: header
      name: x-player-id
      description: Dev-bypass identity (local only, never in production)
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid identity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Pack:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        priceCents: { type: integer }
        currency: { type: string }
        songCount: { type: integer }
        purchasable: { type: boolean }
        owned: { type: boolean }
security:
  - bearerAuth: []
  - devPlayerId: []